Fix usage of the `since` parameter
The streams that used the `since` query parameter for bookmarking were
broken and performing a full sync each run. The `_format_since` function
was copied from the pyxero library:
https://github.com/freakboy3742/pyxero/blob/2ef5e35c0bb3c39977d9eccb79aba8ce41179f1a/xero/basemanager.py#L292-L297
but is not correct, at least in our case. Maybe it works in that library
for some reason specific to the way the library works.

With this change I tested using the `since` parameter with streams:

- bank_transactions
- contacts
- invoices
- overpayments
- prepayments
- accounts
- items
- payments

And confirmed they all work now.